:recycle: refactored App component for navigator implementation
