+++ miam/App.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import MainNavigator from './src/navigators/MainNavigator'
+
+const App = () => {
+    return <MainNavigator/>
+}
+
+export default App
