+++ miam/index.js
@@ -3,8 +3,8 @@
  */
 
 import {AppRegistry} from 'react-native';
-import App from './src/screens/Home/App';
 import {name as appName} from './app.json';
 import 'react-native-gesture-handler';
+import App from './App';
 
 AppRegistry.registerComponent(appName, () => App);
